ML model about everythings that I could do.
This train.ipynp is the way to Manufacture PKL files. The ml.py stores the method that app_ml.py need to use. Except predicting for a certain hour, I also write the function to get the StandsAvailable for an average time and predict for a day, which might be helpful